+++ src/frontend/src/hooks/misc.hooks.ts
@@ -7,7 +7,8 @@ import { useQuery } from 'react-query';
 import { VersionObject } from '../utils/types';
 import { getReleaseInfo } from '../apis/misc.api';
 import { useHistory } from 'react-router-dom';
-import { useState } from 'react';
+import { useContext, useState } from 'react';
+import { ClarityContext } from '../app/ClarityProvider';
 
 export const useGetVersionNumber = () => {
   return useQuery<VersionObject, Error>(['version'], async () => {
@@ -33,4 +34,18 @@ export const useHistoryState = <T>(key: string, initialValue: T): [T, (t: T) => 
     rawSetState(value);
   }
   return [rawState, setState];
+};
+
+/**
+ * useClarity hook
+ *
+ * Returns the Clarity function from context. Use this to call Clarity API methods.
+ * Example: const clarity = useClarity();
+ */
+export const useClarity = () => {
+  const context = useContext(ClarityContext);
+  if (context === undefined) {
+    throw new Error('useClarity must be used within a ClarityProvider');
+  }
+  return context;
 };
