+++ src/frontend/src/tests/pages/WorkPackageDetailPage/WorkPackagePage.test.tsx
@@ -13,13 +13,11 @@ import { mockAuth, mockUseQueryResult } from '../../test-support/test-data/test-
 import { exampleDesignWorkPackage, exampleResearchWorkPackage } from '../../test-support/test-data/work-packages.stub';
 import { exampleWbsProject1 } from '../../test-support/test-data/wbs-numbers.stub';
 import WorkPackagePage from '../../../pages/WorkPackageDetailPage/WorkPackagePage';
-import AppContextUser from '../../../app/AppContextUser';
 import { useCurrentUser } from '../../../hooks/users.hooks';
 import {
   exampleAuthenticatedAdminUser,
   exampleAuthenticatedGuestUser
 } from '../../test-support/test-data/authenticated-user.stub';
-import ClarityProvider from '../../../app/ClarityProvider';
 
 vi.mock('../../../hooks/work-packages.hooks');
 
@@ -55,11 +53,7 @@ const renderComponent = () => {
   const RouterWrapper = routerWrapperBuilder({});
   return render(
     <RouterWrapper>
-      <ClarityProvider>
-        <AppContextUser>
-          <WorkPackagePage wbsNum={exampleWbsProject1} />
-        </AppContextUser>
-      </ClarityProvider>
+      <WorkPackagePage wbsNum={exampleWbsProject1} />
     </RouterWrapper>
   );
 };
